DOMCore/tests/submissions/Ms2ger/Node-insertBefore.html

<!DOCTYPE html>
<title>Node.insertBefore</title>
<script src="http://w3c-test.org/resources/testharness.js"></script>
<div id="log"></div>
<script>
function testHRE(node) {
  test(function() {
    assert_throws("HIERARCHY_REQUEST_ERR", function() { node.insertBefore(document.createTextNode("fail"), null) })
    assert_throws(new TypeError(), function() { node.insertBefore(null, null) })
  })
}
test(function() {
  assert_throws("HIERARCHY_REQUEST_ERR", function() { document.body.insertBefore(document.body, document.getElementById("log")) })
  assert_throws("HIERARCHY_REQUEST_ERR", function() { document.body.insertBefore(document.documentElement, document.getElementById("log")) })

  assert_throws(new TypeError(), function() { document.body.insertBefore(null, null) })

  assert_throws(new TypeError(), function() { document.body.insertBefore(null, document.body.firstChild) })

  assert_throws(new TypeError(), function() { document.body.insertBefore({'a':'b'}, document.body.firstChild) })

  testHRE(document.createTextNode("Foo"))
})
</script>
